Prepare version 0.12.0
Changed

- Enhancement to the file already exists message. Now the collection is
clickable and the document in the collection is highlighted;
- The sidebar on the documents section now list collections in
alphabetical order
- Collections filter improvements
- The collection filters now shows the parent of a collection on mouse
over
- The project icon has been removed and the same color bars are used to
identify project and personal collections
- Available collection filters in a collection are limited to the
current collection and its sub-collections. No more collections from
other projects or personal
- Finally fixed the possibility of selecting collections to be
considered in AND, i.e. to find documents that are both in two or more
collections at the same time
- No more locked (dark grey) collections in filters