Swtich authorization checks to policies and gates
Originally all controllers had a middleware to check for authorization based on a configuration file, route-permissions, that was highly dependent from the route name and sometimes not updated when new routes were created.
Laravel introduced the concepts Policy and Gate to perform authorization on the current user, so we are switching all http controllers to use those.